+++ lib/invoice_item.rb
@@ -27,4 +27,5 @@ class InvoiceItem
 
 
 
+
 end
