+++ lib/item.rb
@@ -19,4 +19,6 @@ class Item
   def unit_price_to_dollars
     (unit_price.to_f / 100).to_f
   end
+
+  
 end
