+++ lib/merchant_repository.rb
@@ -66,4 +66,10 @@ attr_reader :merchant_instances
     all.map {|merchant| merchant.invoices.count}
   end
 
+  def most_sold_item_for_merchant(merchant_id)
+    all.map do |merchant|
+      # binding.pry
+      merchant.most_sold_items
+    end
+  end
 end
